% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethyLiution.R
\name{chooseBestArray}
\alias{chooseBestArray}
\title{Step 4: Where multiple arrays are present for some samples, choose the best to carry forwards to subsequent analyses
        Detect the duplicated assays for the same sample, and keep the one assay with the lowest mean detection p-value
        Remove methylation values with higherdetection p-value (larger than --maxDetectionPval)
        Remove probes with failed detection (p-value> maxDetectionPval) in more than -maxDetectionProp number of samples}
\usage{
chooseBestArray(begin, mset_file, logfile, max_pval, max_prop,
  detection_freq_infofile, detection_filter_infofile, outdir,
  col.name = NULL, mset.lumi = NULL)
}
\arguments{
\item{begin}{Step the pipeline begins at}

\item{mset_file}{Path to file in which dataset object is stored between steps (auto-filled by runPipeline() function)}

\item{logfile}{Path to log file (auto-filled by runPipeline() function)}

\item{max_pval}{Upper threshold for considering a result significant}

\item{max_prop}{For each probe, the freq/proportion of assays that fail by max_pval of detection}

\item{detection_freq_infofile}{Duplicated assays output file (auto-filled by runPipeline() function)}

\item{detection_filter_infofile}{Detection failure table by probe file (auto-filled by runPipeline() function)}

\item{outdir}{Path to output directory}

\item{col.name}{List of column names, generated by setColumnNames() (auto-filled by runPipeline() function)}

\item{mset.lumi}{Microarray dataset object (auto-filled by runPipeline() function)}
}
\value{
mset.lumi dataset, with less performant assays for each sample removed
}
\description{
This probably isn't something you want to run directly. Just use the runPipeline() function.
}
\keyword{internal}
