% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethyLiution.R
\name{cite_me}
\alias{cite_me}
\title{Cite MethyLiution}
\usage{
cite_me()
}
\value{
Citation for this software, in BibTeX format
}
\description{
If this software has proved useful to you, please cite it!
}
\examples{
cite_me()
}
