% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethyLiution.R
\name{colourAdjustment}
\alias{colourAdjustment}
\title{Step 6: Colour balancing for 2-colour microarrays
Applies a normalisation method to remove systemic differences between colour channels.}
\usage{
colourAdjustment(begin, mset_file, logfile, outdir, col.name = NULL,
  mset.lumi = NULL, sub_assays = NULL)
}
\arguments{
\item{begin}{Step the pipeline begins at}

\item{mset_file}{Path to file in which dataset object is stored between steps (auto-filled by runPipeline() function)}

\item{logfile}{Path to log file (auto-filled by runPipeline() function)}

\item{outdir}{Path to output directory}

\item{col.name}{List of column names, generated by setColumnNames() (auto-filled by runPipeline() function)}

\item{mset.lumi}{Microarray dataset object (auto-filled by runPipeline() function)}

\item{sub_assays}{Object containing breakdown of sub-assays, made during Step 5.}
}
\value{
list - mset.lumi dataset, colour-balanced dataset object
}
\description{
This probably isn't something you want to run directly. Just use the runPipeline() function.
}
\keyword{internal}
