% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethyLiution.R
\name{detectOutliers}
\alias{detectOutliers}
\title{Step 5: Detect (but do not remove) outliers
Makes a hierarchical clustering plot of all samples.}
\usage{
detectOutliers(begin, mset_file, outliers_file, logfile, max_plot, outdir,
  col.name = NULL, mset.lumi = NULL)
}
\arguments{
\item{begin}{Step the pipeline begins at}

\item{mset_file}{Path to file in which dataset object is stored between steps (auto-filled by runPipeline() function)}

\item{outliers_file}{Path to outliers output file}

\item{logfile}{Path to log file (auto-filled by runPipeline() function)}

\item{max_plot}{Maximum number of plots to create at this step}

\item{outdir}{Path to output directory}

\item{col.name}{List of column names, generated by setColumnNames() (auto-filled by runPipeline() function)}

\item{mset.lumi}{Microarray dataset object (auto-filled by runPipeline() function)}
}
\value{
list - mset.lumi dataset, breakdown of sub-assays
}
\description{
This probably isn't something you want to run directly. Just use the runPipeline() function.
}
\keyword{internal}
