% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethyLiution.R
\name{findSurrogateVariables}
\alias{findSurrogateVariables}
\title{Step 10: Apply surrogate variable analysis to automatically find hidden variables in the data}
\usage{
findSurrogateVariables(begin, mset_file, sv_file, logfile,
  col.name = NULL, mset.lumi = NULL, bmiq.M = NULL)
}
\arguments{
\item{begin}{Step the pipeline begins at}

\item{mset_file}{Path to file in which dataset object is stored between steps (auto-filled by runPipeline() function)}

\item{sv_file}{Path to file where any surrogate variables found are written as output (auto-filled by runPipeline() function)}

\item{logfile}{Path to log file (auto-filled by runPipeline() function)}

\item{col.name}{List of column names, generated by setColumnNames() (auto-filled by runPipeline() function)}

\item{mset.lumi}{Microarray dataset object (auto-filled by runPipeline() function)}

\item{bmiq.M}{Object containing BMIQ-normalised data, made during Step 8. (auto-filled by runPipeline() function)}
}
\value{
mset.lumi dataset
}
\description{
This probably isn't something you want to run directly. Just use the runPipeline() function.
}
\keyword{internal}
