% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethyLiution.R
\name{normalisation}
\alias{normalisation}
\title{Step 7: Apply inter-assay normalisation
Apply a broader inter-array normalisation procedure than the colour balancing in Step 6, determined by user input.
Options are quantile, SSN, and none.}
\usage{
normalisation(begin, mset_file, logfile, normalisation_method, outdir,
  col.name = NULL, mset.lumi = NULL, lumi.C = NULL,
  sub_assays = NULL)
}
\arguments{
\item{begin}{Step the pipeline begins at}

\item{mset_file}{Path to file in which dataset object is stored between steps (auto-filled by runPipeline() function)}

\item{logfile}{Path to log file (auto-filled by runPipeline() function)}

\item{normalisation_method}{Normalisation method: one of 'quantile', 'ssn', or 'none'.}

\item{outdir}{Path to output directory}

\item{col.name}{List of column names, generated by setColumnNames() (auto-filled by runPipeline() function)}

\item{mset.lumi}{Microarray dataset object (auto-filled by runPipeline() function)}

\item{lumi.C}{Colour-balanced dataset object (auto-filled by runPipeline() function)}

\item{sub_assays}{Object containing breakdown of sub-assays, made during Step 5. (auto-filled by runPipeline() function)}
}
\value{
list - mset.lumi dataset, meth.B normalised dataset object (exprs), meth.M normalised dataset object (betas)
}
\description{
This probably isn't something you want to run directly. Just use the runPipeline() function.
}
\keyword{internal}
