% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethyLiution.R
\docType{data}
\name{probe_data}
\alias{probe_data}
\title{Filtered probe lists for EPIC and 450K datasets}
\format{A list of two data frames.}
\usage{
data(probe_data)
}
\description{
The list of valid probes (for both EPIC and 450K arrays) required for step 2 is included as an accessible dataset, `probe_data`. This dataset was derived from the list of probes available in the BioConductor packages `IlluminaHumanMethylationEPICanno.ilm10b2.hg19` and `IlluminaHumanMethylation450kprobe`, with filtering applied to remove probes located within 10bp of a SNP in order to avoid misleading results arising from polymorphism-induced hybridisation issues.
}
\keyword{datasets}
