% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethyLiution.R
\name{quantileNormalisation}
\alias{quantileNormalisation}
\title{Step 8: Apply Beta-Mixture Quantile normalisation, producing BMIQ objects}
\usage{
quantileNormalisation(begin, mset_file, logfile, BMIQ_dir, outdir, threads,
  col.name = NULL, mset.lumi = NULL, meth.B = NULL)
}
\arguments{
\item{begin}{Step the pipeline begins at}

\item{mset_file}{Path to file in which dataset object is stored between steps (auto-filled by runPipeline() function)}

\item{logfile}{Path to log file (auto-filled by runPipeline() function)}

\item{BMIQ_dir}{Path to a directory where BMIQ normalised data can be written to files (auto-filled by runPipeline() function)}

\item{outdir}{Path to output directory}

\item{threads}{Number of threads to run certain tasks in parallel}

\item{col.name}{List of column names, generated by setColumnNames() (auto-filled by runPipeline() function)}

\item{mset.lumi}{Microarray dataset object (auto-filled by runPipeline() function)}

\item{meth.B}{Object containing normalised data, made during Step 7. (auto-filled by runPipeline() function)}
}
\value{
list - mset.lumi dataset, BMIQ-normalised data object (B values), BMIQ-normalised data object (M-values)
}
\description{
This probably isn't something you want to run directly. Just use the runPipeline() function.
}
\keyword{internal}
