% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethyLiution.R
\name{readData}
\alias{readData}
\title{Step 1: Read dataset files}
\usage{
readData(datadir, metafile, logfile, mset_file, array = "450K",
  threads = 1, FDB)
}
\arguments{
\item{datadir}{Path to directory containing IDAT files}

\item{metafile}{Path to file containing experimental metadata}

\item{logfile}{Path to log file (auto-filled by runPipeline() function)}

\item{mset_file}{Path to file in which dataset object is stored between steps (auto-filled by runPipeline() function)}

\item{array}{Microarray probeset used: valid options are "450K" and "EPIC" (default "450K")}

\item{threads}{Number of threads to run certain tasks in parallel}

\item{FDB}{Methylation annotation set (auto-filled by runPipeline() function)}
}
\value{
List - column names and methylumi dataset
}
\description{
This probably isn't something you want to run directly. Just use the runPipeline() function.
}
\keyword{internal}
