% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethyLiution.R
\name{removeAssaysWithMissingMetadata}
\alias{removeAssaysWithMissingMetadata}
\title{Step 9: Remove assays with missing metadata}
\usage{
removeAssaysWithMissingMetadata(begin, mset_file, logfile,
  feature_data_file, metafile, missing_meta_file, meta_data_file, outdir,
  col.name = NULL, mset.lumi = NULL, meth.B = NULL, meth.M = NULL,
  bmiq.B = NULL, bmiq.M = NULL)
}
\arguments{
\item{begin}{Step the pipeline begins at}

\item{mset_file}{Path to file in which dataset object is stored between steps (auto-filled by runPipeline() function)}

\item{logfile}{Path to log file (auto-filled by runPipeline() function)}

\item{feature_data_file}{Path to a file containing feature data (auto-filled by runPipeline() function)}

\item{metafile}{Path to file containing experimental metadata}

\item{missing_meta_file}{Path to a file where assays with missing data are explicitly listed (auto-filled by runPipeline() function)}

\item{meta_data_file}{Path to a file containing valid metadata from samples that have passed this filter (auto-filled by runPipeline() function)}

\item{outdir}{Path to output directory}

\item{col.name}{List of column names, generated by setColumnNames() (auto-filled by runPipeline() function)}

\item{mset.lumi}{Microarray dataset object (auto-filled by runPipeline() function)}

\item{meth.B}{Object containing normalised data, made during Step 7. (auto-filled by runPipeline() function)}

\item{meth.M}{Object containing normalised data, made during Step 7. (auto-filled by runPipeline() function)}

\item{bmiq.B}{Object containing BMIQ-normalised data, made during Step 8. (auto-filled by runPipeline() function)}

\item{bmiq.M}{Object containing BMIQ-normalised data, made during Step 8. (auto-filled by runPipeline() function)}
}
\value{
list - mset.lumi dataset, BMIQ-normalised data object (B values)
}
\description{
This probably isn't something you want to run directly. Just use the runPipeline() function.
}
\keyword{internal}
