% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethyLiution.R
\name{runPipeline}
\alias{runPipeline}
\title{runPipeline}
\usage{
runPipeline(datadir, metafile, outdir, array = "450K", max_pval = 0.05,
  max_prop = 0.5, max_plot = 30, threads = 1,
  gender_stringent = FALSE, normalisation_method = "quantile",
  begin = 1, end = 9)
}
\arguments{
\item{datadir}{Path to directory containing IDAT files}

\item{metafile}{Path to file containing experimental metadata}

\item{outdir}{Path to output directory}

\item{array}{Microarray probeset used: valid options are "450K" and "EPIC" (default "450K")}

\item{max_pval}{Upper threshold for considering a result significant (default 0.05)}

\item{max_prop}{For each probe, the freq/proportion of assays that fail by max_pval of detection (default 0.5)}

\item{max_plot}{Maximum number of plot output files to create (default 30)}

\item{threads}{Number of threads to run certain tasks in parallel (default 1)}

\item{gender_stringent}{Should the pipeline discard samples where inferred gender is inconsistent with metadata? (default FALSE)}

\item{normalisation_method}{Normalisation method: one of 'quantile', 'ssn', or 'none'. (default 'quantile')}

\item{begin}{Step the pipeline begins at (default 1)}

\item{end}{Step the pipeline ends at (default 9)}
}
\value{
path to output directory
}
\description{
This function is the means of accessing quality control functions of the MethyLiution package. Most procedures are invoked in a set order due to requirements for output from previous steps, so a single function invoking those procedures as a fixed workflow is preferable. \cr
Analysis is broken down into a pipeline of 10 steps: \cr
\itemize{
 \item 1. Read data. \cr
 \item 2. Remove SNP probes and imprinting genes. \cr
 \item 3. Check reported sex against inferred sex. \cr
 \item 4. Remove all but the most performant assay per sample. \cr
 \item 5. Detect (but do not remove) outliers. \cr
 \item 6. Normalise for dye colour. \cr
 \item 7. Apply inter-assay normalisation. \cr
 \item 8. Apply beta-mixture quantile (intra-sample) normalisation. \cr
 \item 9. Remove assays with missing data. \cr
 \item 10. Apply surrogate variable analysis to discover hidden variables. \cr
}
For more details, see vignette(MethyLiution).
}
\examples{
runPipeline()
}
