% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethyLiution.R
\name{setColumnNames}
\alias{setColumnNames}
\title{Set column names}
\usage{
setColumnNames(mset.lumi)
}
\arguments{
\item{mset.lumi}{A methylation dataset object}
}
\value{
New column names
}
\description{
Set column names of input dataset to consistent values.
}
